+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Audit /download_results endpoint support
+- Sanitize service support
+
 ### Changed
 
 - Replaced usage of the deprecated io/ioutil package.
-- Audit /download_results endpoint support
-- Sanitize service support
 
 ### Fixed