+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## v3.8.0-beta.1 - 2024-03-18
+
 ### Changed
 
 - Replaced usage of the deprecated io/ioutil package.